pyproject.toml
[tool.poetry]
name = "booking-service"
version = "0.1.0"
description = ""
authors = ["Dmitry"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.3.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
